fix: update template java imports for 0.4.4-SNAPSHOT package rename
KafkaTopicDescriptor, OwnedKafkaTopicInput, OwnedKafkaTopicOutput and
related types moved to org.creekservice.api.kafka.metadata.topic package
in 0.4.4-SNAPSHOT. Fix template files used by add_service.sh.
