+++ .creek/service_template/services/src/main/java/io/github/creek/service/ks/aggregate/api/demo/services/ExampleServiceDescriptor.java
@@ -23,8 +23,8 @@ import java.util.Collection;
 import java.util.List;
 import io.github.creek.service.ks.aggregate.api.demo.api.KsAggregateApiDemoAggregateDescriptor;
 import io.github.creek.service.ks.aggregate.api.demo.internal.TopicConfigBuilder;
-import org.creekservice.api.kafka.metadata.OwnedKafkaTopicInput;
-import org.creekservice.api.kafka.metadata.OwnedKafkaTopicOutput;
+import org.creekservice.api.kafka.metadata.topic.OwnedKafkaTopicInput;
+import org.creekservice.api.kafka.metadata.topic.OwnedKafkaTopicOutput;
 import org.creekservice.api.platform.metadata.ComponentInput;
 import org.creekservice.api.platform.metadata.ComponentInternal;
 import org.creekservice.api.platform.metadata.ComponentOutput;
